[BSK-1427] Added image for testing and trying to get the module to work. Although the canny transform is working segfaults are occurring
Former-commit-id: 3f61653ec31a3c0a50fabbf4d93f7e17528fca08
Former-commit-id: b25cff5211b71a1ad50d653689cd66b8ba631f3a

+++ src/simulation/simFswInterfaceMessages/limbOpNavMsg.h
@@ -25,7 +25,7 @@ typedef struct {
     int valid; //!< --  Valid measurement if 1, not if 0*/
     int64_t cameraID;          //!< -- [-]   ID of the camera that took the snapshot*/
     double planetIds;          //!< -- [-]   ID for identified celestial body*/
-    double limbPoints[3*MAX_LIMB_PNTS];          //!< -- [-] (x, y, 1) in pixels of the limb points*/
+    int limbPoints[3*MAX_LIMB_PNTS];          //!< -- [-] (x, y, 1) in pixels of the limb points*/
     double pointSigmas[3*MAX_LIMB_PNTS];          //!< -- [-] (sigma_x, _sigma_y, 0) in pixels of the limb points*/
 }LimbOpNavMsg;
